Move db.ts into utils
